requirements.txt
streamlit>=1.31.0
python-dotenv>=1.0.0
pydantic-ai
rich>=13.7.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
streamlit-google-auth>=0.1.0
google-api-python-client>=2.0.0
extra-streamlit-components==0.1.60
pinecone-client>=3.0.0
openai>=1.0.0